+++ lib/internal/streams/pipeline.js
@@ -7,7 +7,6 @@ const {
   ArrayIsArray,
   ReflectApply,
   SymbolAsyncIterator,
-  SymbolIterator,
 } = primordials;
 
 let eos;
@@ -21,6 +20,12 @@ const {
   ERR_MISSING_ARGS,
   ERR_STREAM_DESTROYED
 } = require('internal/errors').codes;
+
+const {
+  isIterable,
+  isReadable,
+  isStream,
+} = require('internal/streams/utils');
 
 let EE;
 let PassThrough;
@@ -76,26 +81,6 @@ function popCallback(streams) {
   if (typeof streams[streams.length - 1] !== 'function')
     throw new ERR_INVALID_CALLBACK(streams[streams.length - 1]);
   return streams.pop();
-}
-
-function isReadable(obj) {
-  return !!(obj && typeof obj.pipe === 'function');
-}
-
-function isWritable(obj) {
-  return !!(obj && typeof obj.write === 'function');
-}
-
-function isStream(obj) {
-  return isReadable(obj) || isWritable(obj);
-}
-
-function isIterable(obj, isAsync) {
-  if (!obj) return false;
-  if (isAsync === true) return typeof obj[SymbolAsyncIterator] === 'function';
-  if (isAsync === false) return typeof obj[SymbolIterator] === 'function';
-  return typeof obj[SymbolAsyncIterator] === 'function' ||
-    typeof obj[SymbolIterator] === 'function';
 }
 
 function makeAsyncIterable(val) {
